refactor: ディレクトリ名を変更
