+++ camera_system/camera_system.py
